+++ packages/core/src/lib/components/image/hooks/image-viewer/index.ts
@@ -1,3 +1,4 @@
 export * from "./use-cursor-visibility";
 export * from "./use-image-navigation";
 export * from "./use-image-scale";
+export * from "./use-prevent-scroll";
